requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
cohere==4.37.0
qdrant-client==2.7.2
tiktoken==0.5.2
nltk==3.8.1
python-dotenv==1.0.0
pyyaml==6.0.1
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
mypy==1.7.1
types-requests==2.31.0.10
types-PyYAML==6.0.12.12
